src/pyproject.toml
[project]
name = "ideologicalatlas-backend"
version = "0.1.0"
description = "Backend API for Ideological Atlas Project"
authors = [
    {name = "Ideological Atlas Team", email = "support@ideologicalatlas.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3==1.42.21",
    "celery==5.6.2",
    "django==6.0",
    "django-axes==8.1.0",
    "django-cities-light==3.10.2",
    "django-cors-headers==4.9.0",
    "django-debug-toolbar==6.1.0",
    "django-environ==0.12.0",
    "django-extensions==4.1",
    "django-filter==25.2",
    "django-model-utils==5.0.0",
    "django-silk==5.4.3",
    "django-simple-history==3.11.0",
    "django-storages==1.14.6",
    "django-unfold==0.75.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-spectacular==0.29.0",
    "flower==2.0.1",
    "gunicorn==23.0.0",
    "ipython==9.8.0",
    "pillow==12.1.0",
    "psycopg==3.3.2",
    "requests==2.32.5",
    "types-requests==2.32.4.20250913",
    "uvicorn==0.40.0",
    "whitenoise==6.11.0",
]

[project.optional-dependencies]
dev = [
    "httpx==0.28.1",
    "pre-commit==4.5.1",
    "coverage==7.13.1",
    "factory-boy==3.3.3",
    "freezegun==1.5.5",
    "tblib==3.0.0",
    "types-pytz==2025.2.0.20251108"
]

[tool.coverage.run]
branch = true
source = ["apps"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/factories/*"
]
concurrency = ["multiprocessing"]
parallel = true

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 90
